Move ApplicationStatus to ApplicationTasksStatus
The tasks status looks better being calculated in
the hook than in the component.  Moved it there
and readjusted the component.  Now the summarized
tasks status is available everywhere a
`DecoratedApplication` is used.
